+++ exchange_rate_scraper/constants.py
